♻️ [Refactoring]: #041 content-stream/ 配下を 1機能1フォルダ化
- graphics-state/ 5 件 (graphics-state, line-cap, line-join, matrix, stack)
- operators/graphics-state/ 7 件 (cm, graphics-state-operators, line-{cap,join,width}-handler, miter-limit-handler, numeric-pdf-object)
- tokenizer/inline-image 1 件
- 各テストを {feature}/__tests__/ へ git mv
- 主実装内の相対 import を +1、test 内を +2 階層調整
- tokenizer.inline-image.test.ts は手動で inline-image/__tests__/ へ移動
- Test Files=124 / Tests=1688 維持
