+++ packages/core/src/content-stream/operators/graphics-state/graphics-state-operators/__tests__/graphics-state-operators.basic.test.ts
@@ -1,0 +1,15 @@
+import { assert, expect, test } from "vitest";
+import { OperatorRegistry } from "../../../../operator-registry/index";
+import { registerGraphicsStateOperators } from "../../graphics-state-operators";
+
+test.each([
+  ["cm"],
+  ["w"],
+  ["J"],
+  ["j"],
+  ["M"],
+])("registerGraphicsStateOperators は %s を登録する", (name) => {
+  const result = registerGraphicsStateOperators(OperatorRegistry.create());
+  assert(result.ok);
+  expect(OperatorRegistry.has(result.value, name)).toBe(true);
+});
